+++ R/zz_arguments.R
@@ -106,6 +106,8 @@
 #'   all available cores will be used. Defaults to "*sequential*".
 #' @param .first_resample A list containing the `.R` resamples based on the original
 #'   data obtained by resamplecSEMResults().
+#' @param .force Logical. Should .object be resampled even if it contains resamples
+#'   already?. Defaults to `FALSE`.
 #' @param .fit_measures Logical. (EXPERIMENTAL) Should additional fit measures 
 #'   be included? Defaults to `FALSE`. 
 #' @param .full_output Logical. Should the full output of summarize be printed.
@@ -123,6 +125,7 @@
 #' @param .id Character string or integer. A character string giving the name or 
 #'   an integer of the position of the column of `.data` whose levels are used
 #'   to split `.data` into groups. Defaults to `NULL`.
+#' @param .inference Logical. Should critical values be computed? Defaults to `FALSE`.
 #' @param .independent Character string. The name of the independent variable. Defaults to `NULL`.
 #' @param .independent_1 Character string. The name of the first independent variable. Defaults to `NULL`.
 #' @param .independent_2 Character string. The name of the second independent variable. Defaults to `NULL`.
@@ -344,11 +347,14 @@ args_csem_dotdotdot <- function(
 #' list shows which argument is passed to which (internal) function:
 #' \describe{
 #' \item{.absolute}{Accepted by/Passed down to: [calculateHTMT()]}
-#' \item{.alpha}{Accepted by/Passed down to: [calculateRhoT()]}
+#' \item{.alpha}{Accepted by/Passed down to: [calculateRhoT()] and [calculateHTMT()]}
 #' \item{.closed_form_ci}{Accepted by/Passed down to: [calculateRhoT()]}
+#' \item{.handle_inadmissibles}{Accepted by/Passed down to: [calculateHTMT()]}
 #' \item{.null_model}{Accepted by/Passed down to: [calculateDf()]}
+#' \item{.R}{Accepted by/Passed down to: [calculateHTMT()]}
 #' \item{.saturated}{Accepted by/Passed down to: [calculateSRMR()], 
 #'   [calculateDG()], [calculateDL()], [calculateDML()]and subsequently [fit()].}
+#' \item{.seed}{Accepted by/Passed down to: [calculateHTMT()]}
 #' \item{.type_vcv}{Accepted by/Passed down to: [calculateSRMR()], 
 #'   [calculateDG()], [calculateDL()], [calculateDML()] and subsequently [fit()].}
 #' }
@@ -361,8 +367,12 @@ args_assess_dotdotdot <- function(
   .absolute            = TRUE,
   .alpha               = 0.05,
   .closed_form_ci      = FALSE,
+  .handle_inadmissibles= c("drop", "ignore", "replace"),
+  .inference           = FALSE,
   .null_model          = FALSE,
+  .R                   = 499,
   .saturated           = FALSE,
+  .seed                = NULL,
   .type_vcv            = "indicator"
 ) {NULL}
   
@@ -420,12 +430,14 @@ args_default <- function(.choices = FALSE) {
     .E                       = NULL,
     .effect                  = NULL,
     .eval_plan               = c("sequential", "multiprocess"),
+    .force                   = FALSE,
     .fit_measures            = FALSE,
     .first_resample          = NULL,
     .full_output             = TRUE,
     .handle_inadmissibles    = c("drop", "ignore", "replace"),
     .H                       = NULL,
     .id                      = NULL,
+    .inference               = FALSE,
     .independent             = NULL,
     .independent_1           = NULL,
     .independent_2           = NULL,